BREAKING(front_matter): move to top-level folder
