+++ encoding/front_matter/json.ts
@@ -1,11 +1,19 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
-import { createExtractor, Format, Parser, test as _test } from "./mod.ts";
-export { Format } from "./mod.ts";
+import { extract } from "../../front_matter/json.ts";
 
-export function test(str: string): boolean {
-  return _test(str, [Format.JSON]);
-}
+export {
+  /** @deprecated (will be removed after 0.182.0) Import from `std/front_matter/json.ts` instead. */
+  Format,
+} from "./mod.ts";
 
-export const extract = createExtractor({ [Format.JSON]: JSON.parse as Parser });
+export {
+  /** @deprecated (will be removed after 0.182.0) Import from `std/front_matter/json.ts` instead. */
+  test,
+} from "../../front_matter/json.ts";
+
+export {
+  /** @deprecated (will be removed after 0.182.0) Import from `std/front_matter/json.ts` instead. */
+  extract,
+};
 export default extract;
