--- encoding/front_matter/testdata/json.md
@@ -1,13 +1,0 @@
----json
-{
-  "title": "Three dashes followed by the format marks the spot",
-  "tags": [
-    "json",
-    "front-matter"
-  ],
-  "expanded-description": "with some ---json 🙃 crazy stuff in it"
-}
----
-don't break
----
-{Also: "---json this shouldn't be a problem"}
