--- encoding/front_matter/testdata/toml.md
@@ -1,8 +1,0 @@
----toml
-title = 'Three dashes followed by the format marks the spot'
-tags = ['toml', 'front-matter']
-'expanded-description' = 'with some ---toml 👌 crazy stuff in it'
----
-don't break
----
-Also = '---toml this shouldn't be a problem'
