+++ encoding/front_matter/toml.ts
@@ -1,13 +1,21 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
-import { createExtractor, Format, Parser, test as _test } from "./mod.ts";
-import { parse } from "../toml.ts";
+import { extract } from "../../front_matter/toml.ts";
 
-export { Format } from "./mod.ts";
+export {
+  /** @deprecated (will be removed after 0.182.0) Import from `std/front_matter/toml.ts` instead. */
+  Format,
+} from "./mod.ts";
 
-export function test(str: string): boolean {
-  return _test(str, [Format.TOML]);
-}
+export {
+  /** @deprecated (will be removed after 0.182.0) Import from `std/front_matter/toml.ts` instead. */
+  test,
+} from "../../front_matter/toml.ts";
 
-export const extract = createExtractor({ [Format.TOML]: parse as Parser });
+export {
+  /** @deprecated (will be removed after 0.182.0) Import from `std/front_matter/toml.ts` instead. */
+  extract,
+};
+
+/** @deprecated (will be removed after 0.182.0) Import from `std/front_matter/toml.ts` instead. */
 export default extract;
