+++ front_matter/_test_utils.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
-import { assert, assertEquals, assertThrows } from "../../testing/asserts.ts";
-import { dirname, fromFileUrl, join, resolve } from "../../path/mod.ts";
+import { assert, assertEquals, assertThrows } from "../testing/asserts.ts";
+import { dirname, fromFileUrl, join, resolve } from "../path/mod.ts";
 import { Format } from "./mod.ts";
 
 const moduleDir = dirname(fromFileUrl(import.meta.url));
@@ -82,6 +82,7 @@ export async function runExtractJSONTests(
   extractFn: ExtractFn,
 ) {
   const str = await Deno.readTextFile(resolveTestDataPath("json.md"));
+  console.log(resolveTestDataPath("json.md"));
   const content = extractFn(str);
 
   assert(content !== undefined);
@@ -115,6 +116,7 @@ export async function runExtractYAMLTests1(
   extractFn: ExtractFn,
 ) {
   const str = await Deno.readTextFile(resolveTestDataPath("yaml1.md"));
+  console.log(resolveTestDataPath("yaml1.md"));
   const content = extractFn(str);
 
   assert(content !== undefined);
