+++ front_matter/mod_test.ts
@@ -1,9 +1,9 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
-import { assert, assertThrows } from "../../testing/asserts.ts";
+import { assert, assertThrows } from "../testing/asserts.ts";
 import { createExtractor, Format, Parser, test } from "./mod.ts";
-import { parse as parseYAML } from "../yaml.ts";
-import { parse as parseTOML } from "../toml.ts";
+import { parse as parseYAML } from "../encoding/yaml.ts";
+import { parse as parseTOML } from "../toml/parse.ts";
 import {
   resolveTestDataPath,
   runExtractJSONTests,
