+++ front_matter/testdata/yaml2.md
@@ -1,0 +1,11 @@
+---yaml title: Three dashes marks the spot tags:
+
+- yaml
+- front-matter
+- dashes expanded-description: with some --- crazy stuff in it
+
+---
+
+## don't break
+
+Also this shouldn't be a problem
